moved the product types into its separate folder
